refactored as per class discussion
