+++ foods.html
@@ -90,7 +90,7 @@
           <thead>
             <tr>
               <th scope="col">Name</th>
-              <th scope="col">Calories</th>
+              <th scope="col" class="calories-header">Calories</th>
               <th scope="col"></th>
             </tr>
           </thead>
