+++ index.html
@@ -26,7 +26,7 @@
             </tr>
           </thead>
           <tbody>
-            <td class="goal"></td>
+            <td class="goal">2000</td>
             <td class="calories-consumed"></td>
             <td class="total-remaining-calories"></td>
           </tbody>
@@ -66,7 +66,7 @@
           <thead>
             <tr>
               <th scope="col">Name</th>
-              <th scope="col" class="calories-header" data-sort-order="default">Calories</th>
+              <th scope="col" class="calories-header">Calories</th>
               <th scope="col"></th>
             </tr>
           </thead>
